Prevent execution unless the processor is operating.

+++ ReplicaConsole/IWindow.cs
@@ -4,7 +4,7 @@ namespace ReplicaConsole;
 
 public interface IWindow
 {
-    public nint WindowId { get; }
+    public uint WindowId { get; }
     public void HandleEvent(SDL.SDL_Event e);
     public void Update();
     public void Close();
